Fill functions.

+++ core/client/Client.cpp
@@ -5,7 +5,7 @@
 namespace core::client {
   using boost::asio::ip::tcp;
 
-  void Client::connect(std::string_view const &address, uint8_t const &port) {
+  void Client::connect(std::string_view const &address, int const &port) {
     socket_ = new boost::asio::ip::tcp::socket(io_service_);
 
     socket_->connect(tcp::endpoint(
